+++ lib/modules/auth/auth_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: auth
-// Updated at: 2026-05-27 17:28:15
-// Commit: feat(auth): add bi-directional event bus handler for auth
+// Updated at: 2026-05-27 18:23:25
+// Commit: refactor(auth): extract reusable atomic widgets from auth
 
 class authComponentState {
-  final String commitRef = '1507';
-  final String lastMessage = 'feat(auth): add bi-directional event bus handler for auth';
-  final DateTime timestamp = DateTime.parse('2026-05-27T17:28:15+05:00');
+  final String commitRef = '1511';
+  final String lastMessage = 'refactor(auth): extract reusable atomic widgets from auth';
+  final DateTime timestamp = DateTime.parse('2026-05-27T18:23:25+05:00');
 }